set a consistent DeletionPolicy for RDS resources based on stack type
Depending on whether it's a dev, stage or prod stack, all resources will
be site to "Retain" or all will be set to "Delete". This controls what
happens to the resources in the event CloudFormation deletes the stack
or replaces the resources. If set to "Retain" things like the db
instances and parameter groups will be kept around but moved out of the
Cfn stack.